TimeUtil: remove redundant casts
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@16217 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/TimeUtil.java
@@ -57,55 +57,55 @@ public class TimeUtil {
             return sb.toString();
         }
         //
-        long weeks = (long)(l / ((long)7*24*60*60*1000));
+        long weeks = (l / ((long)7*24*60*60*1000));
         if (weeks > 0) {
             sb.append(weeks).append('w');
             termCount++;
-            l = l - ((long)weeks * ((long)7*24*60*60*1000));
+            l = l - (weeks * ((long)7*24*60*60*1000));
         }
         if(termCount >= maxTerms) {
             return sb.toString();
         }
         //
-        long days = (long)(l / ((long)24*60*60*1000));
+        long days = (l / ((long)24*60*60*1000));
         if (days > 0) {
             sb.append(days).append('d');
             termCount++;
-            l = l - ((long)days * ((long)24*60*60*1000));
+            l = l - (days * ((long)24*60*60*1000));
         }
         if(termCount >= maxTerms) {
           return sb.toString();
         }
         //
-        long hours = (long)(l / ((long)60*60*1000));
+        long hours = (l / ((long)60*60*1000));
         if (hours > 0) {
             sb.append(hours).append('h');
             termCount++;
-            l = l - ((long)hours * ((long)60*60*1000));
+            l = l - (hours * ((long)60*60*1000));
         }
         if(termCount >= maxTerms) {
             return sb.toString();
         }
         //
-        long minutes = (long)(l / ((long)60*1000));
+        long minutes = (l / ((long)60*1000));
         if (minutes > 0) {
             sb.append(minutes).append('m');
             termCount++;
-            l = l - ((long)minutes * ((long)60*1000));
+            l = l - (minutes * ((long)60*1000));
         }
         if(termCount >= maxTerms) {
             return sb.toString();
         }
         if(withSecondFractions && ((maxTerms - termCount) >= 2)) {
             if (l > 0) {
-                double fractionalSeconds = ((double) l) / ((double) 1000.0);
+                double fractionalSeconds = ((double) l) / (1000.0);
                 DecimalFormat fix3 = new DecimalFormat("0.000");
                 sb.append(fix3.format(fractionalSeconds)).append('s');
                 termCount++;
                 //l = l - ((long)fractionalSeconds * (long)1000);
             }
         } else {
-            long seconds = (long)(l / (long)1000);
+            long seconds = (l / (long)1000);
             if (seconds > 0) {
                 sb.append(seconds).append('s');
                 termCount++;
